Use more appropriate types for MaxUserCount and OperCount.

+++ include/users.h
@@ -23,8 +23,8 @@ typedef Anope::unordered_map<User *> user_map;
 
 extern CoreExport user_map UserListByNick, UserListByUID;
 
-extern CoreExport int OperCount;
-extern CoreExport unsigned MaxUserCount;
+extern CoreExport size_t OperCount;
+extern CoreExport size_t MaxUserCount;
 extern CoreExport time_t MaxUserTime;
 
 /* Online user and channel data. */
